# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CRM Customer Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List all employees
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create a new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeBody"
      responses:
        "201":
          description: Employee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    put:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeBody"
      responses:
        "200":
          description: Employee updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ========================
  # CUSTOMER STATUSES
  # ========================
  /customer-statuses:
    get:
      operationId: listCustomerStatuses
      tags: [settings]
      summary: List customer statuses
      responses:
        "200":
          description: List of statuses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomerStatus"
    post:
      operationId: createCustomerStatus
      tags: [settings]
      summary: Create customer status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerStatusBody"
      responses:
        "201":
          description: Status created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerStatus"
  /customer-statuses/{id}:
    put:
      operationId: updateCustomerStatus
      tags: [settings]
      summary: Update customer status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerStatusBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerStatus"
    delete:
      operationId: deleteCustomerStatus
      tags: [settings]
      summary: Delete customer status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /customer-statuses/reorder:
    post:
      operationId: reorderCustomerStatuses
      tags: [settings]
      summary: Reorder customer statuses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderBody"
      responses:
        "200":
          description: Reordered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ========================
  # CUSTOMER SOURCES
  # ========================
  /customer-sources:
    get:
      operationId: listCustomerSources
      tags: [settings]
      summary: List customer sources
      responses:
        "200":
          description: List of sources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomerSource"
    post:
      operationId: createCustomerSource
      tags: [settings]
      summary: Create customer source
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerSourceBody"
      responses:
        "201":
          description: Source created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerSource"
  /customer-sources/{id}:
    put:
      operationId: updateCustomerSource
      tags: [settings]
      summary: Update customer source
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerSourceBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerSource"
    delete:
      operationId: deleteCustomerSource
      tags: [settings]
      summary: Delete customer source
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ========================
  # PRODUCT TYPES
  # ========================
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers with pagination and filtering
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
        - name: statusId
          in: query
          schema:
            type: integer
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: needFollowUp
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated customers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerListResponse"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    put:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ========================
  # REPORTS
  # ========================
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    ReorderBody:
      type: object
      properties:
        ids:
          type: array
          items:
            type: integer
      required: [ids]

    # ========================
    # EMPLOYEE
    # ========================
    Employee:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        isProtected:
          type: boolean
        createdAt:
          type: string
      required: [id, name, role]
    CreateEmployeeBody:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
      required: [name, role]

    # ========================
    # CUSTOMER STATUS
    # ========================
    CustomerStatus:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
        sortOrder:
          type: integer
        isSystem:
          type: boolean
      required: [id, name, color, sortOrder]
    CreateCustomerStatusBody:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
        sortOrder:
          type: integer
      required: [name, color]

    # ========================
    # CUSTOMER SOURCE
    # ========================
    CustomerSource:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        icon:
          type: string
      required: [id, name]
    CreateCustomerSourceBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        icon:
          type: string
      required: [name]

    # ========================
    # PRODUCT TYPE
    # ========================
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        createdAt:
          type: string
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        note:
          type: string
        statusId:
          type: integer
        statusName:
          type: string
        statusColor:
          type: string
        employeeId:
          type: integer
        employeeName:
          type: string
        sourceId:
          type: integer
        sourceName:
          type: string
        lastContactAt:
          type: string
          nullable: true
        nextContactAt:
          type: string
          nullable: true
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        totalRevenue:
          type: integer
        totalProfit:
          type: integer
      required: [id, name, phone, statusId, employeeId]
    CustomerListResponse:
      type: object
      properties:
        customers:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
      required: [customers, total, page, pageSize, totalPages]
    CreateCustomerBody:
      type: object
      properties:
        createdAt:
          type: string
          nullable: true
        name:
          type: string
        phone:
          type: string
        address:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
        statusId:
          type: integer
        employeeId:
          type: integer
        sourceId:
          type: integer
          nullable: true
        lastContactAt:
          type: string
          nullable: true
        nextContactAt:
          type: string
          nullable: true
        orders:
          type: array
          items:
            $ref: "#/components/schemas/CreateOrderBody"
      required: [name, phone, statusId, employeeId]

    # ========================
    # ORDER
    # ========================
    Order:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        closedAt:
          type: string
          nullable: true
        orderCode:
          type: string
          nullable: true
        productTypeId:
          type: integer
          nullable: true
        productTypeName:
          type: string
          nullable: true
        supplySourceId:
          type: integer
          nullable: true
        supplySourceName:
          type: string
          nullable: true
        productId:
          type: integer
          nullable: true
        productName:
          type: string
          nullable: true
        customProductName:
          type: string
          nullable: true
        revenue:
          type: integer
        profit:
          type: integer
        warrantyMonths:
          type: integer
          nullable: true
        warrantyExpiry:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, customerId, revenue, profit]
    CreateOrderBody:
      type: object
      properties:
        closedAt:
          type: string
          nullable: true
        orderCode:
          type: string
          nullable: true
        productTypeId:
          type: integer
          nullable: true
        supplySourceId:
          type: integer
          nullable: true
        productId:
          type: integer
          nullable: true
        customProductName:
          type: string
          nullable: true
        revenue:
          type: integer
        profit:
          type: integer
        warrantyMonths:
          type: integer
          nullable: true
        note:
          type: string
          nullable: true
      required: [revenue, profit]

    # ========================
    # REPORTS
    # ========================
...