# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CRM Customer Management API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOrderBody {
  closedAt?: string | null;
  orderCode?: string | null;
  productTypeId?: number | null;
  supplySourceId?: number | null;
  productId?: number | null;
  customProductName?: string | null;
  revenue: number;
  profit: number;
  warrantyMonths?: number | null;
  note?: string | null;
}