# lib/api-zod/src/generated/types/customerSource.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CRM Customer Management API
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerSource {
  id: number;
  name: string;
  description?: string;
  icon?: string;
}