# lib/api-zod/src/generated/types/customerStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CRM Customer Management API
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerStatus {
  id: number;
  name: string;
  color: string;
  sortOrder: number;
  isSystem?: boolean;
}