# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CRM Customer Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  id: number;
  customerId: number;
  closedAt?: string | null;
  orderCode?: string | null;
  productTypeId?: number | null;
  productTypeName?: string | null;
  supplySourceId?: number | null;
  supplySourceName?: string | null;
  productId?: number | null;
  productName?: string | null;
  customProductName?: string | null;
  revenue: number;
  profit: number;
  warrantyMonths?: number | null;
  warrantyExpiry?: string | null;
  note?: string | null;
  createdAt?: string;
}